refactor(de): split up de a bit
Move each struct into its own module rather than having them all be in the same one
