+++ src/lib.rs
@@ -29,6 +29,9 @@ mod defs;
 mod map_variant_serializer;
 mod seq_serializer;
 mod map_serializer;
+mod variant_visitor;
+mod ext_visitor;
+mod seq_visitor;
 pub mod error;
 pub mod ser;
 pub mod de;
